# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Associació cultural de Montcada",
        "version": "0.1"
    },
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Display a listing of the roles",
                "description": "Display a listing of the roles.",
                "operationId": "indexRole",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Store a role in database",
                "description": "Store a role in database.",
                "operationId": "storeRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Admin"
                                    },
                                    "guard_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "web"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/role/{roleId}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Display a specific role",
                "description": "Display a specific role.",
                "operationId": "showRole",
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "path",
                        "description": "Id of role that needs to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/roles/{roleId}": {
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Update a role in database",
                "description": "Update a role in database.",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "path",
                        "description": "Id of role that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Admin"
                                    },
                                    "guard_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "web"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Delete a role in database",
                "description": "Delete a role in database.",
                "operationId": "deleteRole",
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "path",
                        "description": "Id of role that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}